warnings
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@6089 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ MyPoint.h
@@ -24,7 +24,7 @@
 }
 
 @property(assign) NSPoint point;
-@property(readonly) CGFloat x, y;
+@property(readonly) float x, y;
 
 +(MyPoint*)point:(NSPoint)a;
 -(id)initWithPoint:(NSPoint)a;
